feat(202-01): extract mergeFieldOverrides into shared helper
Extract mergeFieldOverrides, replaceStringField, and replaceBoolField
from use-agent-save.ts into hooks/helpers/merge-field-overrides.ts.
The shared helper accepts a fieldKeyMap parameter for entity-specific
field mappings. use-agent-save.ts now imports from the shared module.
