+++ .planning/STATE.md
@@ -3,7 +3,7 @@
 ## Current Position
 
 - **Current Milestone:** v8.1.0 — Studio Polish & Prompt Quality
-- **Status:** Planning
+- **Status:** Executing
 - **Task Complexity:** COMPLEX
 - **Oversight:** milestone
 - **Last Updated:** 2026-03-25
@@ -32,7 +32,9 @@
 
 ## Allowed Events
 
-- `PLAN_COMPLETE`
+- `PHASE_COMPLETE`
+- `PHASE_FAILED`
+- `SUSPEND`
 
 ## Previous Milestones
 
@@ -59,4 +61,4 @@ _State updated: 2026-03-25 — v6.0.0 milestone archived, state reset for next m
 
 ---
 
-_State generated from machine snapshot at 2026-03-25T23:29:06.722Z_
+_State generated from machine snapshot at 2026-03-25T23:43:09.381Z_
